requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
werkzeug==2.3.7
python-dotenv==1.0.0